Criada a camada de domínio referente a Anotacao

+++ src/main/java/br/com/totemAutoatendimento/aplicacao/pedido/dto/DadosDePedido.java
@@ -17,7 +17,7 @@ public class DadosDePedido {
 
     private String descricaoDaMercadoria;
 
-    private Integer quantidade;
+    private Float quantidade;
 	
 	private LocalDate data;
